// src/app/model/doublingInstrumentDto.ts
export interface DoublingInstrumentDto {
    id?: string;
    createdBy?: string;
    createdAt?: Date;
    modifiedBy?: string;
    modifiedAt?: Date;
    instrumentId?: string;
    levelAssessmentInner?: number;
    levelAssessmentTeam?: number;
    availabilityId?: string;
    comment?: string;
}